Use Patching Base to define patches compatible older rubies
